large refactor, break up /middleware/ into several files
