+++ src/actions/blocks.js
@@ -3,7 +3,7 @@ import { merge } from 'lodash';
 import * as ActionTypes from '../constants/ActionTypes';
 import BlockDefinition from '../schemas/Block';
 import Block from '../models/Block';
-import { saveBlock, loadBlock } from '../middleware/api';
+import { saveBlock, loadBlock } from '../middleware/data';
 import * as selectors from '../selectors/blocks';
 import * as projectSelectors from '../selectors/projects';
 import * as undoActions from '../store/undo/actions';
