+++ src/middleware/data.js
@@ -1,0 +1,119 @@
+import rejectingFetch from './rejectingFetch';
+import invariant from 'invariant';
+import { headersGet, headersPost, headersPut, headersDelete } from './headers';
+import { dataApiPath } from './paths';
+import ProjectDefinition from '../schemas/Project';
+import BlockDefinition from '../schemas/Block';
+
+/*************************
+ Data API
+ *************************/
+
+export const createBlock = (block, projectId) => {
+  invariant(projectId, 'Project ID is required');
+  invariant(BlockDefinition.validate(block), 'Block does not pass validation: ' + block);
+
+  try {
+    const stringified = JSON.stringify(block);
+    const url = dataApiPath(`${projectId}/${block.id}`);
+
+    return rejectingFetch(url, headersPut(stringified))
+      .then(resp => resp.json());
+  } catch (err) {
+    return Promise.reject('error stringifying block');
+  }
+};
+
+export const loadBlock = (blockId, projectId = 'block') => {
+  invariant(projectId, 'Project ID is required');
+  invariant(blockId, 'Block ID is required');
+
+  const url = dataApiPath(`${projectId}/${blockId}`);
+
+  return rejectingFetch(url, headersGet())
+    .then(resp => resp.json());
+};
+
+//in general, youll probably want to save the whole project? but maybe not.
+export const saveBlock = (block, projectId, overwrite = false) => {
+  invariant(projectId, 'Project ID is required');
+  invariant(BlockDefinition.validate(block), 'Block does not pass validation: ' + block);
+
+  try {
+    const stringified = JSON.stringify(block);
+    const url = dataApiPath(`${projectId}/${block.id}`);
+    const headers = !!overwrite ? headersPut : headersPost;
+
+    return rejectingFetch(url, headers(stringified))
+      .then(resp => resp.json());
+  } catch (err) {
+    return Promise.reject('error stringifying block');
+  }
+};
+
+//returns metadata of projects
+export const listProjects = () => {
+  const url = dataApiPath('projects');
+  return rejectingFetch(url, headersGet())
+    .then(resp => resp.json())
+    .then(projects => projects.filter(project => !!project));
+};
+
+//saves just the project manifest to file system
+export const saveProjectManifest = (project) => {
+  invariant(ProjectDefinition.validate(project), 'Project does not pass validation: ' + project);
+
+  try {
+    const stringified = JSON.stringify(project);
+    const url = dataApiPath(`${project.id}`);
+
+    return rejectingFetch(url, headersPost(stringified))
+      .then(resp => resp.json());
+  } catch (err) {
+    return Promise.reject('error stringifying project');
+  }
+};
+
+//returns a rollup
+export const loadProject = (projectId) => {
+  const url = dataApiPath(`projects/${projectId}`);
+  return rejectingFetch(url, headersGet())
+    .then(resp => resp.json());
+};
+
+//expects a rollup
+//autosave
+//returns the commit with sha, message
+export const saveProject = (projectId, rollup) => {
+  invariant(projectId, 'Project ID required to snapshot');
+  invariant(rollup, 'Rollup is required to save');
+  invariant(rollup.project && Array.isArray(rollup.blocks), 'rollup in wrong form');
+
+  const url = dataApiPath(`projects/${projectId}`);
+  const stringified = JSON.stringify(rollup);
+
+  return rejectingFetch(url, headersPost(stringified))
+    .then(resp => resp.json());
+};
+
+//expects a rollup
+//explicit, makes a git commit with special message to differentiate
+//returns the commit wth sha, message
+export const snapshot = (projectId, rollup, message = 'Project Snapshot') => {
+  invariant(projectId, 'Project ID required to snapshot');
+  invariant(!message || typeof message === 'string', 'optional message for snapshot must be a string');
+
+  const stringified = JSON.stringify({ message });
+  const url = dataApiPath(`${projectId}/commit`);
+
+  return rejectingFetch(url, headersPost(stringified))
+    .then(resp => resp.json());
+};
+
+//todo - make this more explicit, this is a bit low-level
+export const infoQuery = (type, detail) => {
+  const url = dataApiPath(`info/${type}${detail ? `/${detail}` : ''}`);
+  return rejectingFetch(url, headersGet())
+    .then(resp => resp.json());
+};
+
