+++ test/middleware/extensions.js
@@ -1,12 +1,12 @@
 import chai from 'chai';
 import fs from 'fs';
-import * as api from '../../src/middleware/api';
+import * as extensionsApi from '../../src/middleware/extensions';
 const { assert, expect } = chai;
 
 describe('Middleware', () => {
   describe('Extensions', () => {
     it('getExtensionsInfo() should be able get extension manifests', () => {
-      return api.getExtensionsInfo()
+      return extensionsApi.getExtensionsInfo()
         .then(output => {
           assert(typeof output === 'object', 'wrong format for info');
           assert(Object.keys(output).every(key => {
